misc: - rename decklayers - fix 100%map under MUI toggle drawer
