+++ src/views/CityScopeJS/BaseMap/deckglLayers/AggregatedTripsLayer.js
