+++ src/views/CityScopeJS/BaseMap/index.js
@@ -1,7 +1,6 @@
 import React, { Component } from "react";
 import BaseMap from "./BaseMap";
 import { connect } from "react-redux";
-import { ProjectionMapping } from "../ProjectionMapping/ProjectionMapping";
 
 class MapContainer extends Component {
     _checkKeystone = () => {
@@ -10,26 +9,10 @@ class MapContainer extends Component {
 
     render() {
         return (
-            <div
-                style={{
-                    height: "100vh",
-                    width: "100vw",
-                    overflow: "hidden",
-                }}
-            >
-                <ProjectionMapping
-                    style={{
-                        height: "100vh",
-                        width: "100vw",
-                    }}
-                    isEditMode={this._checkKeystone()}
-                >
-                    <BaseMap
-                        menu={this.props.menu}
-                        selectedType={this.props.selectedType}
-                    />
-                </ProjectionMapping>
-            </div>
+            <BaseMap
+                menu={this.props.menu}
+                selectedType={this.props.selectedType}
+            />
         );
     }
 }
